% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Proteinminion-functions.R
\name{CustomDB_Fisher}
\alias{CustomDB_Fisher}
\title{CustomDB_Fisher()}
\usage{
CustomDB_Fisher(query, universe, custom_db, id_colname, ontology_colname)
}
\arguments{
\item{query}{should be a character vector corresponding to a query list (UniProt_Accession or UniProt_ID)}

\item{universe}{should be a character vector corresponding to a universe list (Uniprot Entry), if missing the all table from the UniProtTable will be used as universe}

\item{custom_db}{should be a data.frame containing at least 2 columns one with the identifiers matching to the query/universe identifiers and one containing the associated ontology terms/groups. For each ID multiple ontology terms can be associated, multiple rows containing the same ID and different ontology terms should be present.}

\item{id_colname}{character vector corresponding to the colname of the column to use for the ids}

\item{ontology_colname}{}
}
\value{
a data.frame containing enrichment results
}
\description{
Allows to perform enrichment analysis from a custom database using a query list, a universe list and an annotation table (containing the identifiers and associated ontology terms)
}
\examples{
UniProtGOfisher(QueryExample, UniverseExample, organism.id=organismExample)

}
