% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Proteinminion-functions.R
\name{download_UniProt_Fasta}
\alias{download_UniProt_Fasta}
\title{download_UniProt_Fasta()}
\usage{
download_UniProt_Fasta(
  organismID = "9606",
  proteomeID = "proteomeID",
  reviewed = FALSE,
  export = FALSE,
  file = "UniProtFasta_organismID_date.fasta"
)
}
\arguments{
\item{organismID}{Should be a UniProt organism ID}

\item{reviewed}{this indicate if the proteins extracted have to be reviewed or not (typically the ones present on swissProt), by default FALSE.}

\item{export}{indicates if the table should be exported for future use (using a previously downloaded table dramatically reduces the analysis time).}

\item{file}{should be a string indicating the location and name of the destination file.}

\item{proteome}{Should be a UniProt proteome ID}
}
\value{
This function returns a data.frame named UniProtFasta containing the detail on the proteins for the queried organism, if the export is set on TRUE, the function will also return the table in the fasta file.
}
\description{
Retrieve UniProt fasta file for a given organism
}
\author{
Geremy Clair
}
