% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Proteinminion-functions.R
\name{download_UniProt_Table}
\alias{download_UniProt_Table}
\title{download_UniProt_Table()}
\usage{
download_UniProt_Table(
  organismID = "9606",
  proteomeID = "proteomeID",
  reviewed = FALSE,
  export = FALSE,
  file = "UniProtTable_organismID_date.rda"
)
}
\arguments{
\item{organismID}{Should be a UniProt organism ID, by default Homo sapiens will be used ("9606").}

\item{reviewed}{indicate if the proteins extracted have to be reviewed or not (typically the ones present on swissProt), by default TRUE (the non reviewed proteins are discarded), if FALSE both reviewed and not reviewed will be conserved.}

\item{export}{indicates if the table should be exported for future use (using a previously downloaded table, this reduces the analysis time by limiting data fetching).}

\item{file}{should be a string indicating the location and name of the destination file.}
}
\value{
This function returns a data.frame named UniProtTable containing the detail on the proteins for the queried organism, if the export is set on TRUE, the function will also save the table in the .rda format at the specified location
}
\description{
Retrieve UniProt info for the proteins of a given organism.
}
\author{
Geremy Clair
}
