% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Proteinminion-functions.R
\name{enrichment_filter}
\alias{enrichment_filter}
\title{enrichment_filter()}
\usage{
enrichment_filter(
  enrichment_table,
  Test_p_type = "p",
  Test_p = 0.05,
  foldchange = 1,
  category = c("GO_BP", "GO_CC", "GO_MF", "KEGG_pathway", "REACTOME"),
  min_feature = 1
)
}
\arguments{
\item{enrichment_table}{should be an enrichment table generated with Protein MiniOn.}

\item{Test_p_type}{should be 'p' or 'padj'. By default set to 'p' and no filtering is done by Test_pue.}

\item{Test_p}{should be a numerical value indicative of the maximum p to use for the filtering. 0.05 by default.}

\item{foldchange}{indicate a fold change filtering value for the table. 1 by default}

\item{category}{character vector indicates the categories to keep}

\item{min_feature}{numerical value that indicates the minimum of features in the query to be considered. 1 by default.}

\item{order}{should be TRUE or FALSE indicating if the order of the table should be changed.}
}
\value{
a data.frame containing enrichment results from the KEGG retrieve from UniProt.
}
\description{
Function that allows to display filter enrichment results table generated with Protein MiniOn by Test_pue and fold change.
}
\author{
Geremy Clair
}
