% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Proteinminion-functions.R
\name{enrichment_plot}
\alias{enrichment_plot}
\title{enrichment_plot()}
\usage{
enrichment_plot(enrichment_table, Test_p_type = "p")
}
\arguments{
\item{enrichment_table}{should be an enrichment table generated with Protein MiniOn}

\item{Test_p_type}{should be 'p' or 'padj'. By default set to 'p' and no filtering is done by Test_pue.}
}
\value{
a data.frame containing enrichment results from the KEGG retrieve from UniProt.
}
\description{
Function that allows to display barplot of enrichment results table generated with Protein MiniOn
}
\examples{
UniProt_KEGG_EASE(QueryExample, UniverseExample, organism.id=organismExample)
}
\author{
Geremy Clair
}
