% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Proteinminion-functions.R
\name{enrplot}
\alias{enrplot}
\title{enrplot()}
\usage{
enrplot(enrichment_table, pval_type = "pval")
}
\arguments{
\item{enrichment_table}{should be an enrichment table generated with Protein MiniOn}

\item{pval_type}{should be 'p' or 'padj'. By default set to 'p' and no filtering is done by pvalue.}
}
\value{
a data.frame containing enrichment results from the KEGG retrieve from UniProt.
}
\description{
Function that allows to display barplot of enrichment results table generated with Protein MiniOn
}
\examples{
UniProt_KEGG_EASE(QueryExample, UniverseExample, organism.id=organismExample)

}
