% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Proteinminion-functions.R
\name{extractIDs}
\alias{extractIDs}
\title{extractIDs()}
\usage{
extractIDs(term, term_type)
}
\arguments{
\item{term}{should be a character vector of length 1 containing the term accession or description}

\item{term_type}{should be a term_type in c('GO', 'KEGG', 'REACTOME')}
}
\value{
a data.frame containing enrichment results from the KEGG retrieve from UniProt.
}
\description{
Function that allows to extract all the proteins contained in the UniProtTable corresponding to a certain term type (e.g.UniProtTable_GO, UniProtTable_KEGG, UniProtTable_REACTOME)
}
\examples{
UniProt_KEGG_EASE(QueryExample, UniverseExample, organism.id=organismExample)

}
