% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Proteinminion-functions.R
\name{generate_KEGG_Table}
\alias{generate_KEGG_Table}
\title{generate_KEGG_Table()}
\usage{
generate_KEGG_Table(
  organismID = "9606",
  reviewed = TRUE,
  preloaded_UniProtTable = TRUE,
  export = FALSE,
  file = "UniProtTable_KEGG_organismID_date.rda"
)
}
\arguments{
\item{organismID}{Should be a UniProt organism ID, by default Homo sapiens will be used ("9606")}

\item{reviewed}{this indicate if the proteins extracted have to be reviewed or not (typically the ones present on swissProt), by default TRUE (the non reviewed proteins are discarded)}

\item{preloaded_UniProtTable}{this indicate whether or not the preloaded UniProtTable and UniProtTable_GO should be used, if FALSE the UniProtTable and UniProtTable_GO will be generated again}

\item{export}{indicates if the table should be exported for future use (using a previously downloaded table dramatically reduces the analysis time).}

\item{file}{should be a string indicating the location and name of the destination file it will be saved in *.rda format.}
}
\value{
This function generate a KEGG enrichment table named UniProtTable_KEGG usable with the enrichment functions Uniprot_KEGG_*. If the export is set on, the UniProtTable_KEGG will also be saved in the .rda format at the specified location.
}
\description{
Retrieve UniProt KEGG pathways for the proteins of a given organism
}
\author{
Geremy Clair
}
