% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Proteinminion-functions.R
\name{generate_Ontology_Tables}
\alias{generate_Ontology_Tables}
\title{generate_Ontology_Tables()}
\usage{
generate_Ontology_Tables(
  type = c("GO", "KEGG", "REACTOME"),
  organismID = "9606",
  proteomeID = "UP000005640",
  reviewed = TRUE,
  preloaded_UniProtTable = TRUE
)
}
\arguments{
\item{type}{should be a vector containing the type of uniProtTables to be generated can contain 'GO', 'KEGG', and 'REACTOME')}

\item{organismID}{Should be a UniProt organism ID, by default Homo sapiens will be used ("9606")}

\item{proteomeID}{Should be a proteome accession number from UniProt}

\item{reviewed}{this indicate if the proteins extracted have to be reviewed or not (typically the ones present on swissProt), by default TRUE (the non reviewed proteins are discarded)}

\item{preloaded_UniProtTable}{this indicate wether or not the preloaded UniProtTable and UniProtTable_GO should be used, if FALSE the UniProtTable and UniProtTable_GO will be generated again}
}
\value{
This function generate the Ontology Tables requested.
}
\description{
Retrieve UniProt table and generate the uniProtTables GO, KEGG, and REACTOME if requested
}
\author{
Geremy Clair
}
