% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Proteinminion-functions.R
\name{generate_UniProtTable_GO}
\alias{generate_UniProtTable_GO}
\title{generate_UniProtTable_GO()}
\usage{
generate_UniProtTable_GO(
  organismID = "9606",
  reviewed = TRUE,
  preloaded_UniProtTable = TRUE,
  export = FALSE,
  file = "UniProtTable_GO_organismID_date.rda"
)
}
\arguments{
\item{organismID}{Should be a UniProt organism ID, by default Homo sapiens will be used ("9606")}

\item{reviewed}{this indicate if the proteins extracted have to be reviewed or not (typically the ones present on swissProt), by default TRUE (the non reviewed proteins are discarded)}

\item{preloaded_UniProtTable}{this indicate wether or not the preloaded UniProtTable and UniProtTable_GO should be used, if FALSE the UniProtTable and UniProtTable_GO will be generated again}
}
\value{
This file return a data.frame named UniProtTable_GO containing the list of GOs associated with the different proteins from a UniProtTable. If the export is set on, the UniProtTable_Go will also be saved in the .rda format at the specified location.
}
\description{
Retrieve UniProt GOterms for the proteins of a given organism
}
\examples{
generate_UniProtTable_GO(organismID="9606",reviewed=TRUE,preloaded_UniProtTable=TRUE,export=TRUE,file="UniProtTable_GO_organismID_date.rda")

}
