% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Proteinminion-functions.R
\name{generate_UniProtTable_REACTOME}
\alias{generate_UniProtTable_REACTOME}
\title{generate_UniProtTable_REACTOME()}
\usage{
generate_UniProtTable_REACTOME(
  organismID = "9606",
  reviewed = TRUE,
  preloaded_UniProtTable = TRUE,
  export = FALSE,
  file = "UniProtTable_Reactome_organismID_date.rda"
)
}
\arguments{
\item{organismID}{Should be a UniProt organism ID, by default Homo sapiens will be used ("9606")}

\item{reviewed}{this indicate if the proteins extracted have to be reviewed or not (typically the ones present on swissProt), by default TRUE (the non reviewed proteins are discarded)}

\item{preloaded_UniProtTable}{this indicate wether or not the preloaded UniProtTable and UniProtTable_GO should be used, if FALSE the UniProtTable and UniProtTable_GO will be generated again}
}
\value{
This function generate a reactome table
}
\description{
Retrieve UniProt Reactome pathways for the proteins of a given organism
}
