% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Proteinminion-functions.R
\name{GO_details}
\alias{GO_details}
\title{GO_details()}
\usage{
GO_details()
}
\value{
a df containing the following details : GO.ID, Description, and GO.ontology.type
}
\description{
Retrieve GO_details from the GO.db from Bioconductor
}
\examples{
GO_details<-GO_details()
GO_details
}
