% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_split_fasta.R
\name{splitFasta}
\alias{splitFasta}
\title{splitFasta()}
\usage{
splitFasta(file = "file.fasta", seq_number = 10000)
}
\arguments{
\item{file}{should be a string indicating the location and name of the destination file, by default the file will be named with the format "UniProtTable_organismID_date.txt".}

\item{organismID}{Should be a UniProt organism ID, by default Homo sapiens will be used ("9606").}

\item{reviewed}{indicate if the proteins extracted have to be reviewed or not (typically the ones present on swissProt), by default TRUE (the non reviewed proteins are discarded), if FALSE both reviewed and not reviewed will be conserved.}

\item{export}{indicates if the table should be exported for future use (using a previously downloaded table dramatically reduces the analysis time).}
}
\value{
This function returns a data.frame named UniProtTable containing the detail on the proteins for the queried organism, if the export is set on TRUE, the function will also save the table in the .rda format at the specified location
}
\description{
split a fasta in a sequence of fasta files containing a set number of
}
\examples{
download_UniProtTable(organismID="9606", reviewed=TRUE, export=FALSE, file="UniProtTable_organismID_date.txt")

}
