% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Proteinminion-functions.R
\name{Uniprot_Fasta_Parse}
\alias{Uniprot_Fasta_Parse}
\title{Uniprot_Fasta_Parse()}
\usage{
Uniprot_Fasta_Parse(file = "file.fasta")
}
\arguments{
\item{fasta.file}{should be a character vector corresponding to a query list (UniProt_Accession or UniProt_ID)}
}
\value{
parses the fasta files into a table containing the different information on the proteins
}
\description{
allow to extract the information from a fasta file in a table format
}
\author{
Geremy Clair
}
