% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Proteinminion-functions.R
\name{UniprotFastaParser}
\alias{UniprotFastaParser}
\title{UniprotFastaParser()}
\usage{
UniprotFastaParser(file = "file.fasta")
}
\arguments{
\item{fasta.file}{should be a character vector corresponding to a query list (UniProt_Accession or UniProt_ID)}
}
\value{
a data.frame containing enrichment results
}
\description{
allow to extract the information from a fasta file in a table format
}
\examples{
UniProtGOfisher(QueryExample, UniverseExample, organism.id=organismExample)

}
